Apply scroller to the contact list UI #48
