+++ client/src/components/ChatSideBanner/ChatSideBanner.tsx
@@ -8,6 +8,7 @@ import AvatarDisplay from '../AvatarDisplay/AvatarDisplay';
 import Search from '../Search/Search';
 import AuthMenu from '../AuthMenu/AuthMenu';
 import Scroller from '../Scroll/Scrollcomponent';
+import ContactsTab from '../ContactsTab/ContactsTab';
 
 interface Props {
   loggedInUser: User;
@@ -43,7 +44,7 @@ const ChatSideBanner = ({ loggedInUser }: Props): JSX.Element => {
         <Search search={search} handleChange={handleChange} />
       </Box>
       <Box>
-        <Scroller />
+        <ContactsTab />
       </Box>
     </Grid>
   );
